Move the implementation of the code generation SPI
Moves the knowledge of which APIs in the Kernel API that the generated
code should call to the main cypher module, so that the 2.3 version of
the language is not tied to a specific version of the internals of Neo4j
