+++ community/cypher/cypher-compiler-2.3/LICENSES.txt
@@ -256,7 +256,6 @@ SUCH DAMAGE.
 
 ------------------------------------------------------------------------------
 BSD License
-  ASM Core
   Scala Compiler
 ------------------------------------------------------------------------------
