+++ community/cypher/cypher-compiler-2.3/NOTICE.txt
@@ -35,6 +35,5 @@ BSD - Scala License
   Scala Library
 
 BSD License
-  ASM Core
   Scala Compiler
