+++ community/cypher/cypher-compiler-2.3/pom.xml
@@ -146,11 +146,6 @@
       <artifactId>neo4j-graph-algo</artifactId>
       <scope>provided</scope>
     </dependency>
-    <dependency>
-      <groupId>org.neo4j</groupId>
-      <artifactId>neo4j-codegen</artifactId>
-      <version>${project.version}</version>
-    </dependency>
 
     <!-- neo4j testing -->
