+++ community/cypher/cypher/pom.xml
@@ -137,6 +137,11 @@
       <groupId>org.neo4j</groupId>
       <artifactId>neo4j-graph-algo</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.neo4j</groupId>
+      <artifactId>neo4j-codegen</artifactId>
+      <version>${project.version}</version>
+    </dependency>
 
     <!-- neo4j testing -->
