+++ community/cypher/cypher/src/test/scala/org/neo4j/cypher/CartesianProductNotificationAcceptanceTest.scala
@@ -26,6 +26,7 @@ import org.neo4j.cypher.internal.compiler.v2_3._
 import org.neo4j.cypher.internal.compiler.v2_3.notification.CartesianProductNotification
 import org.neo4j.cypher.internal.compiler.v2_3.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.compiler.v2_3.tracing.rewriters.RewriterStepSequencer
+import org.neo4j.cypher.internal.spi.v2_3.GeneratedQueryStructure
 import org.neo4j.helpers.Clock
 import org.neo4j.logging.NullLog
 
@@ -87,6 +88,7 @@ class CartesianProductNotificationAcceptanceTest extends CypherFunSuite with Gra
   private def createCompiler() =
     CypherCompilerFactory.costBasedCompiler(
       graph, 128, 0.5, 1000L, Clock.SYSTEM_CLOCK,
+      GeneratedQueryStructure,
       new WrappedMonitors2_3(kernelMonitors),
       new StringInfoLogger2_3(NullLog.getInstance),
       plannerName = Some(GreedyPlannerName),
