+++ community/cypher/cypher/src/test/scala/org/neo4j/cypher/internal/compiler/v2_3/codegen/CodeGeneratorTest.scala
@@ -32,6 +32,7 @@ import org.neo4j.cypher.internal.compiler.v2_3.planner.logical.plans._
 import org.neo4j.cypher.internal.compiler.v2_3.planner.{LogicalPlanningTestSupport, SemanticTable}
 import org.neo4j.cypher.internal.compiler.v2_3.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.compiler.v2_3.{CostBasedPlannerName, NormalMode, ParameterNotFoundException, TaskCloser}
+import org.neo4j.cypher.internal.spi.v2_3.GeneratedQueryStructure
 import org.neo4j.graphdb.Result.{ResultRow, ResultVisitor}
 import org.neo4j.graphdb.{Relationship, Direction, GraphDatabaseService, Node}
 import org.neo4j.helpers.Clock
@@ -43,7 +44,7 @@ import scala.collection.JavaConverters
 
 class CodeGeneratorTest extends CypherFunSuite with LogicalPlanningTestSupport {
 
-  private val generator = new CodeGenerator()
+  private val generator = new CodeGenerator(GeneratedQueryStructure)
 
   test("all nodes scan") { // MATCH a RETURN a
     //given
