+++ community/cypher/cypher/src/test/scala/org/neo4j/cypher/internal/compiler/v2_3/codegen/ir/CompiledProfilingTest.scala
